requirements.txt
Django==2.1.7
djangorestframework==3.9.2
django-ckeditor==5.6.1
django-heroku
gunicorn
